pyproject.toml
[project]
name = "autodraft-clean"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.66.0",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "rank-bm25>=0.2.2",
    "einops>=0.8.2",
    "kiwipiepy>=0.23.0",
]
